feat: Try using original upload filename before inventing a new one
The code now even handles the very unlikely event of a collision when
adding the uuid postfix to the filename.

Yes. It is more likely to be struck by lightning than to have a
collision when we add a shortuuid, but it will allow using a simpler
or shorter postfix in the future.